+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   * `service_token`
   * `rest_api`
 
+* The `/1.1/lines` resource can now create the underlying SIP endpoint during it's creation
+
 * The following resource has been added:
 
   * POST `/2.0/users`